refactor: move test into right package
In order to avoid dublication of tests (in flaky test detection) and align
with src packages move test classes in right locations
