+++ search/search-client-opensearch/src/test/java/io/camunda/search/os/transformers/RangeQueryTransformerTest.java
@@ -5,21 +5,20 @@
  * Licensed under the Camunda License 1.0. You may not use this file
  * except in compliance with the Camunda License 1.0.
  */
-package io.camunda.data.transformers;
+package io.camunda.search.os.transformers;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
-import co.elastic.clients.elasticsearch._types.query_dsl.RangeQuery;
 import io.camunda.search.clients.query.SearchQueryBuilders;
 import io.camunda.search.clients.query.SearchRangeQuery;
-import io.camunda.search.es.transformers.ElasticsearchTransformers;
 import io.camunda.search.transformers.SearchTransfomer;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
+import org.opensearch.client.opensearch._types.query_dsl.RangeQuery;
 
 public class RangeQueryTransformerTest {
 
-  private final ElasticsearchTransformers transformers = new ElasticsearchTransformers();
+  private final OpensearchTransformers transformers = new OpensearchTransformers();
   private SearchTransfomer<SearchRangeQuery, RangeQuery> transformer;
 
   @BeforeEach
